avances en inserciones de prueba en cada entidad

+++ Trabajo Entregable/mongo-test/src/main/java/com/example/mongotest/Controller/ProductController.java
@@ -35,7 +35,7 @@ public class ProductController {
     }
 
     @DeleteMapping("/{id}")
-    void deletePerson(@PathVariable Long id) {
+    void deletePerson(@PathVariable Integer id) {
         repo.deleteById(id);
     }
