+++ Trabajo Entregable/mongo-test/src/main/java/com/example/mongotest/Repository/ProductRepository.java
@@ -3,5 +3,5 @@ package com.example.mongotest.Repository;
 import com.example.mongotest.Model.Product;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface ProductRepository extends MongoRepository<Product,Long> {
+public interface ProductRepository extends MongoRepository<Product,Integer> {
 }
