+++ Trabajo Entregable/mongo-test/src/main/java/com/example/mongotest/Repository/SaleRepository.java
@@ -3,5 +3,5 @@ package com.example.mongotest.Repository;
 import com.example.mongotest.Model.Sale;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface SaleRepository extends MongoRepository<Sale,Long> {
+public interface SaleRepository extends MongoRepository<Sale,Integer> {
 }
